+++ create_ips_tables.sql
@@ -179,7 +179,7 @@ IF OBJECT_ID(N'ips_savedproposal', N'U') IS NULL
 BEGIN
     CREATE TABLE ips_savedproposal (
         id                    bigint        IDENTITY(1,1) NOT NULL,
-        user_id               bigint                      NOT NULL,
+        user_id               int                         NOT NULL,
         label                 nvarchar(200)               NOT NULL,
         data                  nvarchar(max)               NOT NULL,
         risk_profile_override nvarchar(100)               NOT NULL DEFAULT '',
